fix: do only one read in PdGetPortState
PdEnterAlertInterrupt, PdGetPortState, and PdExitAlertInterrupt are
executed consecutively. Before this change, we did a Pd data status
reading in PdGetPortState and one in PdExitAlertInterrupt. In theory, two
consecutive reads should not cause any issue, but we found that the
second read could only get 'FFFFFFFFFFFF' in rdata. Therefore, this
change removes the read in PdExitAlertInterrupt, and moves the
if statement of checking the HPD_IRQ bit in DATA_STATUS to PdGetPortState.

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>