Platform/O6: define BT/WL_RADIO_DISABLE_L gpios in the ACPI table
The issue is that OS doesn't control the GPIOs unless they are defined
in the ACPI table.

This change defines the GPIO pins of WL_RADIO_DISABLE_L (GPIO019) and
BT_RADIO_DISABLE_L (GPIO020)[0]. It's found that when a WLAN/BT module
is attached to the PCIe E Key port, pulling up the GPIO020 and GPIO019
will disable the functions of WiFi and Bluetooth; therefore, they need
to be pulled up to enable the functions. The corresponding iomux
definitions are also defined in Iomux.asl.

The command `gpioinfo` should contain the following information after
this change:
```
"GPIO019" "PRP0001:0d"  output  active-high [used bias-disabled]
"GPIO020" "PRP0001:0e"  output  active-high [used bias-disabled]
```

This change is based on this commit[1].

[0]: https://dl.radxa.com/orion/o6/hw/radxa_orion_o6_v1.20_schematic.pdf
[1]: https://github.com/radxa/edk2-platforms/commit/7e61f64a334dc5b5f8383c90445a569e0801ccb0

Signed-off-by: Zhaoming Luo <luozhaoming@radxa.com>